web_frame/requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
async-generator==1.10
attrs==22.2.0
bcrypt==3.2.2
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.5
cryptography==39.0.2
exceptiongroup
execnet==1.9.0
h11==0.13.0
idna==3.4
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
paramiko==3.1.0
pip==21.3.1
pluggy==1.0.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
pytest==7.0.1
pytest-xdist==3.0.2
PyYAML==6.0
selenium==4.0.0a7
setuptools==59.6.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.15
wheel==0.37.1
wsproto==1.0.0
